Extract `SeedFork` into own file
