+++ src/world/octave_noise.js
@@ -1,21 +1,7 @@
-import { rangeList } from "../utils.js";
 import SimplexNoise from "../libs/simplex-noise/simplex-noise.js";
 
+import { SeedFork } from "./seed.js";
 
-export class SeedFork {
-  static getSeedExtra(what, index) {
-    return `.!${what}[${index}]`
-  }
-
-  static getSeed(orig, what, index) {
-    return String(orig) + this.getSeedExtra(what, index);
-  }
-
-  static getSeeds(seed, what, n) {
-    // get `n` seeds from a single seed
-    return rangeList(n).map(i => this.getSeed(seed, what, i))
-  }
-}
 
 /**
  * @typedef {[number, number, number]} Vec3
