+++ src/world/seed.js
@@ -1,0 +1,17 @@
+import { rangeList } from "../utils.js";
+
+
+export class SeedFork {
+  static getSeedExtra(what, index) {
+    return `.!${what}[${index}]`
+  }
+
+  static getSeed(orig, what, index) {
+    return String(orig) + this.getSeedExtra(what, index);
+  }
+
+  static getSeeds(seed, what, n) {
+    // get `n` seeds from a single seed
+    return rangeList(n).map(i => this.getSeed(seed, what, i))
+  }
+}
