+++ textures/NOTE.txt
@@ -7,5 +7,5 @@ grass textures from:
 stone texture from:
   https://static.wikia.nocookie.net/minecraft_gamepedia/images/d/dc/Stone_%28texture%29_JE5_BE3.png/revision/latest?cb=20201001141805
 
-lots of textures:
+lots of textures from:
   https://minecraft.fandom.com/wiki/List_of_block_textures
